+++ Directory.Packages.props
@@ -11,17 +11,17 @@
     <PackageVersion Include="NUnit" Version="3.13.3" />
     <PackageVersion Include="FsUnit" Version="5.2.0" />
     <PackageVersion Include="NUnit3TestAdapter" Version="4.5.0" />
-    <PackageVersion Include="Avalonia" Version="11.1.0-rc1" />
-    <PackageVersion Include="Avalonia.Themes.Fluent" Version="11.1.0-rc1" />
-    <PackageVersion Include="Avalonia.Android" Version="11.1.0-rc1" />
-    <PackageVersion Include="Avalonia.iOS" Version="11.1.0-rc1" />
-    <PackageVersion Include="Avalonia.Desktop" Version="11.1.0-rc1" />
-    <PackageVersion Include="Avalonia.Browser" Version="11.1.0-rc1" />
-    <PackageVersion Include="Avalonia.Skia" Version="11.1.0-rc1" />
-    <PackageVersion Include="Avalonia.Diagnostics" Version="11.1.0-rc1" />
-    <PackageVersion Include="Avalonia.Controls.ItemsRepeater" Version="11.1.0-rc1" />
-    <PackageVersion Include="Avalonia.Controls.ColorPicker" Version="11.1.0-rc1" />
-    <PackageVersion Include="Avalonia.Controls.DataGrid" Version="11.1.0-rc1" />
+    <PackageVersion Include="Avalonia" Version="11.1.0-rc2" />
+    <PackageVersion Include="Avalonia.Themes.Fluent" Version="11.1.0-rc2" />
+    <PackageVersion Include="Avalonia.Android" Version="11.1.0-rc2" />
+    <PackageVersion Include="Avalonia.iOS" Version="11.1.0-rc2" />
+    <PackageVersion Include="Avalonia.Desktop" Version="11.1.0-rc2" />
+    <PackageVersion Include="Avalonia.Browser" Version="11.1.0-rc2" />
+    <PackageVersion Include="Avalonia.Skia" Version="11.1.0-rc2" />
+    <PackageVersion Include="Avalonia.Diagnostics" Version="11.1.0-rc2" />
+    <PackageVersion Include="Avalonia.Controls.ItemsRepeater" Version="11.1.0-rc2" />
+    <PackageVersion Include="Avalonia.Controls.ColorPicker" Version="11.1.0-rc2" />
+    <PackageVersion Include="Avalonia.Controls.DataGrid" Version="11.1.0-rc2" />
     <PackageVersion Include="Avalonia.Controls.TreeDataGrid" Version="11.0.2" />
   </ItemGroup>
 </Project>
