+++ extensions/Fabulous.Avalonia.ColorPicker/Fabulous.Avalonia.ColorPicker.fsproj
@@ -29,7 +29,7 @@
         <Compile Include="ColorPreviewer.fs" />
     </ItemGroup>
     <ItemGroup>
-        <PackageReference Include="Avalonia.Controls.ColorPicker" VersionOverride="11.1.0-rc1" />
+        <PackageReference Include="Avalonia.Controls.ColorPicker" VersionOverride="11.1.0-rc2" />
         <PackageReference Include="Microsoft.SourceLink.GitHub" PrivateAssets="All" />
     </ItemGroup>
     <ItemGroup>
