+++ extensions/Fabulous.Avalonia.DataGrid/Fabulous.Avalonia.DataGrid.fsproj
@@ -30,7 +30,7 @@
         <Compile Include="DataGrid.fs" />
     </ItemGroup>
     <ItemGroup>
-        <PackageReference Include="Avalonia.Controls.DataGrid" VersionOverride="11.1.0-rc1" />
+        <PackageReference Include="Avalonia.Controls.DataGrid" VersionOverride="11.1.0-rc2" />
         <PackageReference Include="Microsoft.SourceLink.GitHub" PrivateAssets="All" />
     </ItemGroup>
     <ItemGroup>
