+++ extensions/Fabulous.Avalonia.Diagnostics/Fabulous.Avalonia.Diagnostics.fsproj
@@ -28,7 +28,7 @@
         <Compile Include="DevTools.fs" />
     </ItemGroup>
     <ItemGroup>
-        <PackageReference Include="Avalonia.Diagnostics" VersionOverride="11.1.0-rc1">
+        <PackageReference Include="Avalonia.Diagnostics" VersionOverride="11.1.0-rc2">
             <PrivateAssets>build; native; contentfiles; analyzers</PrivateAssets>
         </PackageReference>
         <PackageReference Include="Microsoft.SourceLink.GitHub" PrivateAssets="All" />
