+++ global.json
@@ -1,6 +1,6 @@
 {
   "sdk": {
-    "version": "8.0.200",
+    "version": "8.0.302",
     "rollForward": "latestMajor",
     "allowPrerelease": false
   }
