--- samples/RenderDemo.Browser/AppBundle/index.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<!--  Licensed to the .NET Foundation under one or more agreements. -->
-<!-- The .NET Foundation licenses this file to you under the MIT license. -->
-<html>
-
-<head>
-    <title>Fabulous RenderDemo</title>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="modulepreload" href="./main.js" />
-    <link rel="modulepreload" href="./dotnet.js" />
-    <link rel="modulepreload" href="./avalonia.js" />
-    <link rel="stylesheet" href="./app.css" />
-</head>
-
-<body style="margin: 0px">
-<div id="out">
-    <div id="avalonia-splash">
-        <div class="center">
-            <h2>Powered by</h2>
-            <a class="navbar-brand" href="https://www.avaloniaui.net/" target="_blank">
-                <img src="Logo.svg" alt="Avalonia Logo" width="30" height="24" />
-                Avalonia
-            </a>
-        </div>
-    </div>
-</div>
-<script type='module' src="./main.js"></script>
-</body>
-
-</html>
