--- samples/RenderDemo.Browser/AppBundle/main.js
@@ -1,17 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-import { dotnet } from './dotnet.js'
-// import { dotnet } from './_framework/dotnet.js' // NET8.0
-
-const is_browser = typeof window != "undefined";
-if (!is_browser) throw new Error(`Expected to be running in a browser`);
-
-const dotnetRuntime = await dotnet
-    .withDiagnosticTracing(false)
-    .withApplicationArgumentsFromQuery()
-    .create();
-
-const config = dotnetRuntime.getConfig();
-
-await dotnetRuntime.runMain(config.mainAssemblyName, [globalThis.location.href]);
