+++ samples/RenderDemo.Browser/RenderDemo.Browser.fsproj
@@ -1,20 +1,23 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <RuntimeIdentifier>browser-wasm</RuntimeIdentifier>
-    <WasmMainJSPath>AppBundle\main.js</WasmMainJSPath>
-    <OutputType>Exe</OutputType>
-  </PropertyGroup>
+﻿<Project Sdk="Microsoft.NET.Sdk.WebAssembly">
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net8.0-browser</TargetFramework>
+        <RuntimeIdentifier>browser-wasm</RuntimeIdentifier>
+        <WasmMainJSPath>wwwroot\main.js</WasmMainJSPath>
+        <WasmRuntimeAssetsLocation>./_framework</WasmRuntimeAssetsLocation>
+        <WasmEnableThreads>false</WasmEnableThreads>
+        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+        <DebuggerSupport>true</DebuggerSupport>
+        <WasmDebugLevel>5</WasmDebugLevel>
+    </PropertyGroup>
 
-	<ItemGroup>
-		<None Include="Properties\**" />
-		<None Include="AppBundle\**" />
-		<WasmExtraFilesToDeploy Include="AppBundle\**" />
-		<Compile Include="Program.fs" />
-		<Content Include="runtimeconfig.template.json" />
-	</ItemGroup>
-	<ItemGroup>
-		<PackageReference Include="Avalonia.Browser" />
-		<ProjectReference Include="..\RenderDemo\RenderDemo.fsproj" />
-	</ItemGroup>
+    <ItemGroup>
+        <WasmExtraFilesToDeploy Include="wwwroot\**"/>
+        <None Include="Properties\**"/>
+        <Compile Include="Program.fs"/>
+    </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="Avalonia.Browser"/>
+        <ProjectReference Include="..\RenderDemo\RenderDemo.fsproj"/>
+    </ItemGroup>
 </Project>
