+++ samples/RenderDemo.Browser/wwwroot/favicon.ico
