+++ samples/RenderDemo.Browser/wwwroot/index.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title>RenderDemo.Browser</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <base href="/" />
+    <link rel="modulepreload" href="./main.js" />
+    <link rel="modulepreload" href="./_framework/dotnet.js" />
+    <link rel="modulepreload" href="./_framework/avalonia.js" />
+    <link rel="stylesheet" href="./app.css" />
+</head>
+
+<body style="margin: 0px; overflow: hidden">
+<div id="out">
+    <div id="avalonia-splash">
+        <div class="center">
+            <h2 class="purple">
+                Powered by
+                <a class="highlight" href="https://www.avaloniaui.net/" target="_blank">Avalonia UI</a>
+            </h2>
+        </div>
+        <img class="icon" src="Logo.svg" alt="Avalonia Logo" />
+    </div>
+</div>
+<script type='module' src="./main.js"></script>
+</body>
+
+</html>
